requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers==2.2.2
pypdf2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
scholarly==1.7.11
arxiv==1.4.8
together==0.2.7
groq==0.4.1
faiss-cpu==1.7.4
tiktoken==0.5.2
numpy==1.24.3
pandas==2.0.3